Update UI layouts and link player to UI in Level scene
Adjusted various RectTransform and layout settings in MainMenuUI.prefab and UI.prefab for improved UI alignment and scaling. Added references to MainMenu and animator in MainMenuUI. In Level.unity, linked the player GameObject to the UI prefab and set up references for CharacterStats and playerInventory. Also updated Credits_Animation keyframes and reduced availableSP in UI prefab.